consolidate dns-account-01 integration tests
